Changed NC supplemental rate to 2019 rate; updated NC standard deductions to 10k for single/married and 15k for hoh; implemented rounding 2019 NC tax results
