+++ tests/Unit/Countries/US/NorthCarolina/NorthCarolinaIncome/V20180101/NorthCarolinaIncomeTest.php
@@ -1,0 +1,120 @@
+<?php
+
+namespace Appleton\Taxes\Countries\US\NorthCarolina\NorthCarolinaIncome;
+
+use Appleton\Taxes\Models\Countries\US\NorthCarolina\NorthCarolinaIncomeTaxInformation;
+use Carbon\Carbon;
+
+class NorthCarolinaIncomeTest extends \TestCase
+{
+    public function setUp()
+    {
+        parent::setUp();
+
+        Carbon::setTestNow(
+            Carbon::parse('January 1, 2018 8am', 'America/Chicago')->setTimezone('UTC')
+        );
+    }
+
+    public function testNorthCarolinaIncome()
+    {
+        $results = $this->taxes->calculate(function ($taxes) {
+            $taxes->setHomeLocation($this->getLocation('us.north_carolina'));
+            $taxes->setWorkLocation($this->getLocation('us.north_carolina'));
+            $taxes->setUser($this->user);
+            $taxes->setEarnings(66.68);
+            $taxes->setPayPeriods(260);
+        });
+
+        $this->assertSame(1.81, $results->getTax(NorthCarolinaIncome::class));
+    }
+
+    public function testNorthCarolinaAdditionalWithholding()
+    {
+        NorthCarolinaIncomeTaxInformation::forUser($this->user)->update(['additional_withholding' => 10]);
+
+        $results = $this->taxes->calculate(function ($taxes) {
+            $taxes->setHomeLocation($this->getLocation('us.north_carolina'));
+            $taxes->setWorkLocation($this->getLocation('us.north_carolina'));
+            $taxes->setUser($this->user);
+            $taxes->setEarnings(0);
+            $taxes->setPayPeriods(260);
+        });
+
+        $this->assertSame(0.0, $results->getTax(NorthCarolinaIncome::class));
+
+        $results = $this->taxes->calculate(function ($taxes) {
+            $taxes->setHomeLocation($this->getLocation('us.north_carolina'));
+            $taxes->setWorkLocation($this->getLocation('us.north_carolina'));
+            $taxes->setUser($this->user);
+            $taxes->setEarnings(66.68);
+            $taxes->setPayPeriods(260);
+        });
+
+        $this->assertSame(11.81, $results->getTax(NorthCarolinaIncome::class));
+    }
+
+    public function testNorthCarolinaSupplemental()
+    {
+        $results = $this->taxes->calculate(function ($taxes) {
+            $taxes->setHomeLocation($this->getLocation('us.north_carolina'));
+            $taxes->setWorkLocation($this->getLocation('us.north_carolina'));
+            $taxes->setUser($this->user);
+            $taxes->setEarnings(100);
+            $taxes->setSupplementalEarnings(100);
+        });
+
+        $this->assertSame(5.59, $results->getTax(NorthCarolinaIncome::class));
+    }
+
+    public function testNorthCarolinaIncomeNonNegative()
+    {
+        $results = $this->taxes->calculate(function ($taxes) {
+            $taxes->setHomeLocation($this->getLocation('us.north_carolina'));
+            $taxes->setWorkLocation($this->getLocation('us.north_carolina'));
+            $taxes->setUser($this->user);
+            $taxes->setEarnings(10);
+            $taxes->setPayPeriods(260);
+        });
+
+        $this->assertSame(0.00, $results->getTax(NorthCarolinaIncome::class));
+    }
+
+    public function testNorthCarolinaIncomeUseDefault()
+    {
+        NorthCarolinaIncomeTaxInformation::forUser($this->user)->delete();
+
+        $results = $this->taxes->calculate(function ($taxes) {
+            $taxes->setHomeLocation($this->getLocation('us.north_carolina'));
+            $taxes->setWorkLocation($this->getLocation('us.north_carolina'));
+            $taxes->setUser($this->user);
+            $taxes->setEarnings(0);
+            $taxes->setPayPeriods(260);
+        });
+
+        $this->assertSame(0.0, $results->getTax(NorthCarolinaIncome::class));
+
+        $results = $this->taxes->calculate(function ($taxes) {
+            $taxes->setHomeLocation($this->getLocation('us.north_carolina'));
+            $taxes->setWorkLocation($this->getLocation('us.north_carolina'));
+            $taxes->setUser($this->user);
+            $taxes->setEarnings(66.68);
+            $taxes->setPayPeriods(260);
+        });
+
+        $this->assertSame(1.81, $results->getTax(NorthCarolinaIncome::class));
+    }
+
+    public function testNorthCarolinaIncomeWorkInTennessee()
+    {
+        $results = $this->taxes->calculate(function ($taxes) {
+            $taxes->setHomeLocation($this->getLocation('us.north_carolina'));
+            $taxes->setWorkLocation($this->getLocation('us.tennessee'));
+            $taxes->setUser($this->user);
+            $taxes->setEarnings(66.68);
+            $taxes->setPayPeriods(260);
+        });
+
+        $this->assertSame(1.81, $results->getTax(NorthCarolinaIncome::class));
+    }
+}
